!win: uv_uptime to actuality return a double
Ref: https://github.com/libuv/libuv/issues/1294

+++ src/unix/sunos.c
@@ -601,7 +601,7 @@ int uv_uptime(double* uptime) {
     *uptime = -1;
   } else {
     knp = (kstat_named_t*)  kstat_data_lookup(ksp, (char*) "clk_intr");
-    *uptime = knp->value.ul / hz;
+    *uptime = knp->value.ul / ((double)hz);
   }
   kstat_close(kc);
